Fix missing order trade relation (#881)
* Add market order field
- If this is true we expect futuresOrder to be empty

* Remove comments

* Make sure futuresOrder and marketOrder gets set when handleDelayedOrderRemoved is triggered

* Make margin transfers a bit cleaner

* Move out liquidation to it's own file

* Update tests

* marginDelta can just be added to traders margin

* Remove log

* Not using receipt: true
- even though I think that would be a really nice solution

* Remove accidental commit
