+++ v2/perps-v2/perps-subgraph/src/futures.ts
@@ -1,32 +1,30 @@
-import { Address, BigInt, DataSourceContext, ethereum } from '@graphprotocol/graph-ts';
-import { PerpetualFuturesMarket } from '../generated/templates';
+import { Address, BigInt, DataSourceContext } from '@graphprotocol/graph-ts';
 import {
   DelayedOrderRemoved as DelayedOrderRemovedEvent,
   DelayedOrderSubmitted as DelayedOrderSubmittedEvent,
   FundingRecomputed as FundingRecomputedEvent,
-  PositionLiquidated as PositionLiquidatedEventLegacy,
-  PositionLiquidated1 as PositionLiquidatedEvent,
   MarginTransferred as MarginTransferredEvent,
   MarketAdded as MarketAddedEvent,
   MarketRemoved as MarketRemovedEvent,
   PerpsTracking as PerpsTrackingEvent,
-  PositionFlagged1 as PositionFlaggedEventOld,
   PositionFlagged as PositionFlaggedEventNew,
+  PositionFlagged1 as PositionFlaggedEventOld,
 } from '../generated/FuturesMarketManagerNew/PerpsV2Proxy';
 import {
-  PositionLiquidated,
-  Trader,
-  Synthetix,
-  FuturesPosition,
-  FuturesOrder,
-  FuturesTrade,
+  Frontend,
   FundingRateUpdate,
   FuturesMarginTransfer,
   FuturesMarket,
-  Frontend,
+  FuturesOrder,
+  FuturesPosition,
+  FuturesTrade,
   PositionFlagged,
+  Synthetix,
+  Trader,
 } from '../generated/schema';
+import { PerpetualFuturesMarket } from '../generated/templates';
 import { updateFeeStats } from './historical-trade-stats';
+export * from './liquidation';
 export { handlePositionModified } from './position-modified';
 
 export function handleFuturesMarketAdded(event: MarketAddedEvent): void {
@@ -74,153 +72,6 @@ export function handlePositionFlaggedOld(event: PositionFlaggedEventOld): void {
   positionFlaggedEntity.save();
 }
 
-export function handlePositionLiquidatedLegacy(event: PositionLiquidatedEventLegacy): void {
-  const futuresPositionId = event.address.toHex() + '-' + event.params.id.toHex();
-  const positionLiquidatedEntity = new PositionLiquidated(event.params.id.toString());
-  positionLiquidatedEntity.trader = event.params.account.toHex();
-  positionLiquidatedEntity.market = event.address.toHex();
-  positionLiquidatedEntity.liquidator = event.params.liquidator;
-  positionLiquidatedEntity.size = event.params.size;
-  positionLiquidatedEntity.price = event.params.price;
-  positionLiquidatedEntity.fee = event.params.fee;
-  positionLiquidatedEntity.futuresPosition = futuresPositionId;
-  positionLiquidatedEntity.timestamp = event.block.timestamp;
-  positionLiquidatedEntity.txHash = event.transaction.hash.toHex();
-  positionLiquidatedEntity.save();
-
-  const futuresPosition = FuturesPosition.load(futuresPositionId);
-
-  let feeForSynthetix = BigInt.fromI32(0);
-
-  if (futuresPosition && futuresPosition.margin.gt(BigInt.fromI32(1000).pow(18))) {
-    feeForSynthetix = event.params.fee.minus(BigInt.fromI32(1002).pow(18));
-  } else {
-    const ninetyPercent = event.params.fee
-      .times(BigInt.fromI32(90).pow(18))
-      .div(BigInt.fromI32(100).pow(18));
-    feeForSynthetix = event.params.fee.minus(ninetyPercent);
-  }
-
-  if (futuresPosition) {
-    futuresPosition.isLiquidated = true;
-    futuresPosition.isOpen = false;
-    futuresPosition.closeTimestamp = event.block.timestamp;
-    futuresPosition.feesPaidToSynthetix = futuresPosition.feesPaidToSynthetix.plus(feeForSynthetix);
-    futuresPosition.exitPrice = event.params.price;
-    futuresPosition.save();
-  }
-
-  const tradeEntity = FuturesTrade.load(
-    event.transaction.hash.toHex() + '-' + event.logIndex.minus(BigInt.fromI32(1)).toString()
-  );
-  if (tradeEntity) {
-    tradeEntity.size = event.params.size.times(BigInt.fromI32(-1));
-    tradeEntity.positionClosed = true;
-    tradeEntity.timestamp = event.block.timestamp;
-    tradeEntity.trader = event.params.account.toHex();
-    tradeEntity.futuresPosition = futuresPositionId;
-    tradeEntity.price = event.params.price;
-    tradeEntity.txHash = event.transaction.hash.toHex();
-    tradeEntity.positionSize = BigInt.fromI32(0);
-    tradeEntity.feesPaidToSynthetix = tradeEntity.feesPaidToSynthetix.plus(feeForSynthetix);
-
-    // TODO make sure this is correct. I think we need to subtract something else.
-    tradeEntity.realizedPnl = tradeEntity.realizedPnl
-      .minus(event.params.fee)
-      .minus(tradeEntity.netFunding); // you loose funding when liquidated
-    tradeEntity.netFunding = BigInt.fromI32(0);
-    tradeEntity.type = 'Liquidated';
-    tradeEntity.save();
-  }
-  updateFeeStats(feeForSynthetix, event.address, event.block.timestamp);
-
-  const synthetix = Synthetix.load('synthetix');
-  if (synthetix) {
-    synthetix.feesByLiquidations = synthetix.feesByLiquidations.plus(feeForSynthetix);
-    synthetix.totalLiquidations = synthetix.totalLiquidations.plus(BigInt.fromI32(1));
-    synthetix.save();
-  }
-
-  let trader = Trader.load(event.params.account.toHex());
-  if (trader) {
-    trader.feesPaidToSynthetix = trader.feesPaidToSynthetix.plus(feeForSynthetix);
-    trader.totalLiquidations = trader.totalLiquidations.plus(BigInt.fromI32(1));
-    trader.totalMarginLiquidated = trader.totalMarginLiquidated.plus(event.params.size);
-    trader.margin = trader.margin.minus(event.params.size);
-    trader.save();
-  }
-}
-
-export function handlePositionLiquidated(event: PositionLiquidatedEvent): void {
-  const futuresPositionId = event.address.toHex() + '-' + event.params.id.toHex();
-  const positionLiquidatedEntity = new PositionLiquidated(event.params.id.toString());
-  positionLiquidatedEntity.trader = event.params.account.toHex();
-  positionLiquidatedEntity.market = event.address.toHex();
-  positionLiquidatedEntity.liquidator = event.params.liquidator;
-  positionLiquidatedEntity.size = event.params.size;
-  positionLiquidatedEntity.price = event.params.price;
-  positionLiquidatedEntity.fee = event.params.stakersFee.plus(
-    event.params.flaggerFee.plus(event.params.liquidatorFee)
-  );
-  positionLiquidatedEntity.futuresPosition = futuresPositionId;
-  positionLiquidatedEntity.timestamp = event.block.timestamp;
-  positionLiquidatedEntity.txHash = event.transaction.hash.toHex();
-  positionLiquidatedEntity.save();
-
-  const tradeEntity = FuturesTrade.load(
-    event.transaction.hash.toHex() + '-' + event.logIndex.minus(BigInt.fromI32(1)).toString()
-  );
-  if (tradeEntity) {
-    tradeEntity.size = event.params.size.times(BigInt.fromI32(-1));
-    tradeEntity.positionClosed = true;
-    tradeEntity.timestamp = event.block.timestamp;
-    tradeEntity.trader = event.params.account.toHex();
-    tradeEntity.futuresPosition = futuresPositionId;
-    tradeEntity.price = event.params.price;
-    tradeEntity.txHash = event.transaction.hash.toHex();
-    tradeEntity.positionSize = BigInt.fromI32(0);
-    tradeEntity.feesPaidToSynthetix = tradeEntity.feesPaidToSynthetix.plus(event.params.stakersFee);
-
-    // TODO make sure this is correct. I think we need to subtract something else.
-    tradeEntity.realizedPnl = tradeEntity.realizedPnl
-      .minus(event.params.stakersFee.plus(event.params.liquidatorFee).plus(event.params.flaggerFee))
-      .minus(tradeEntity.netFunding); // you loose funding when liquidated
-    tradeEntity.netFunding = BigInt.fromI32(0);
-    tradeEntity.type = 'Liquidated';
-    tradeEntity.save();
-  }
-
-  updateFeeStats(event.params.stakersFee, event.address, event.block.timestamp);
-
-  const synthetix = Synthetix.load('synthetix');
-  if (synthetix) {
-    synthetix.feesByLiquidations = synthetix.feesByLiquidations.plus(event.params.stakersFee);
-    synthetix.totalLiquidations = synthetix.totalLiquidations.plus(BigInt.fromI32(1));
-    synthetix.save();
-  }
-
-  const futuresPosition = FuturesPosition.load(futuresPositionId);
-  if (futuresPosition) {
-    futuresPosition.isLiquidated = true;
-    futuresPosition.isOpen = false;
-    futuresPosition.closeTimestamp = event.block.timestamp;
-    futuresPosition.feesPaidToSynthetix = futuresPosition.feesPaidToSynthetix.plus(
-      event.params.stakersFee
-    );
-    futuresPosition.exitPrice = event.params.price;
-    futuresPosition.save();
-  }
-
-  let trader = Trader.load(event.params.account.toHex());
-  if (trader) {
-    trader.feesPaidToSynthetix = trader.feesPaidToSynthetix.plus(event.params.stakersFee);
-    trader.totalLiquidations = trader.totalLiquidations.plus(BigInt.fromI32(1));
-    trader.totalMarginLiquidated = trader.totalMarginLiquidated.plus(event.params.size);
-    trader.margin = trader.margin.minus(event.params.size);
-    trader.save();
-  }
-}
-
 export function handleMarginTransferred(event: MarginTransferredEvent): void {
   let trader = Trader.load(event.params.account.toHex());
   let synthetix = Synthetix.load('synthetix');
@@ -256,11 +107,7 @@ export function handleMarginTransferred(event: MarginTransferredEvent): void {
     trader.realizedPnl = BigInt.fromI32(0);
     synthetix.totalTraders = synthetix.totalTraders.plus(BigInt.fromI32(0));
   } else {
-    if (event.params.marginDelta.gt(BigInt.fromI32(0))) {
-      trader.margin = trader.margin.plus(event.params.marginDelta);
-    } else if (event.params.marginDelta.lt(BigInt.fromI32(0))) {
-      trader.margin = trader.margin.minus(event.params.marginDelta);
-    }
+    trader.margin = trader.margin.plus(event.params.marginDelta);
   }
   marginTransferEntity.save();
   synthetix.save();
@@ -278,15 +125,24 @@ export function handleDelayedOrderRemoved(event: DelayedOrderRemovedEvent): void
 
   // Update FuturesOrderEntity
   if (futuresOrderEntity) {
-    updateFeeStats(event.params.keeperDeposit, event.address, event.block.timestamp);
-
     futuresOrderEntity.fee = event.params.keeperDeposit;
     futuresOrderEntity.keeper = event.transaction.from;
-
     if (tradeEntity) {
-      futuresOrderEntity.futuresPosition = tradeEntity.futuresPosition;
-      futuresOrderEntity.status = 'Filled';
+      // All fee data is based on trade entity, only update keeper fee if trade is created
+      updateFeeStats(event.params.keeperDeposit, event.address, event.block.timestamp);
 
+      // Trade entity updates
+      tradeEntity.futuresOrder = futuresOrderEntity.id;
+      tradeEntity.marketOrder = false;
+      // add fee if not self-executed
+      if (futuresOrderEntity.keeper.toString() != futuresOrderEntity.trader) {
+        tradeEntity.feesPaidToSynthetix = tradeEntity.feesPaidToSynthetix.plus(
+          event.params.keeperDeposit
+        );
+      }
+      tradeEntity.save();
+
+      // Update FuturesPosition
       let positionEntity = FuturesPosition.load(tradeEntity.futuresPosition);
       if (positionEntity) {
         positionEntity.feesPaidToSynthetix = positionEntity.feesPaidToSynthetix.plus(
@@ -294,13 +150,6 @@ export function handleDelayedOrderRemoved(event: DelayedOrderRemovedEvent): void
         );
         positionEntity.realizedPnl = positionEntity.realizedPnl.minus(event.params.keeperDeposit);
         positionEntity.save();
-      }
-      // add fee if not self-executed
-      if (futuresOrderEntity.keeper.toString() != futuresOrderEntity.trader) {
-        tradeEntity.feesPaidToSynthetix = tradeEntity.feesPaidToSynthetix.plus(
-          event.params.keeperDeposit
-        );
-        tradeEntity.save();
       }
 
       // Update Synthetix values
@@ -316,6 +165,9 @@ export function handleDelayedOrderRemoved(event: DelayedOrderRemovedEvent): void
         trader.feesPaidToSynthetix = trader.feesPaidToSynthetix.plus(event.params.keeperDeposit);
         trader.save();
       }
+      // Update FuturesOrderEntity
+      futuresOrderEntity.futuresPosition = tradeEntity.futuresPosition;
+      futuresOrderEntity.status = 'Filled';
       futuresOrderEntity.txHash = event.transaction.hash.toHex();
       futuresOrderEntity.save();
     } else {
