+++ v2/perps-v2/perps-subgraph/subgraph.base.yaml
@@ -3,46 +3,6 @@ schema:
   file: ./schema.graphql
 
 templates:
-  # - name: AggregatorProxy
-  #   kind: ethereum/contract
-  #   source:
-  #     abi: AggregatorProxy
-  #   mapping:
-  #     kind: ethereum/events
-  #     apiVersion: 0.0.7
-  #     language: wasm/assemblyscript
-  #     file: ./src/latest-rates.ts
-  #     entities:
-  #       - Candle
-  #       - LatestRate
-  #       - RateUpdate
-  #     abis:
-  #       - name: AggregatorProxy
-  #         file: ./abis/AggregatorProxy.json
-  #     eventHandlers:
-  #       - event: 'AggregatorConfirmed(indexed address,indexed address)'
-  #         handler: 'handleAggregatorProxyAddressUpdated'
-
-  # - name: Aggregator
-  #   kind: ethereum/contract
-  #   source:
-  #     abi: Aggregator
-  #   mapping:
-  #     kind: ethereum/events
-  #     apiVersion: 0.0.7
-  #     language: wasm/assemblyscript
-  #     file: ./src/latest-rates.ts
-  #     entities:
-  #       - Candle
-  #       - LatestRate
-  #       - RateUpdate
-  #     abis:
-  #       - name: Aggregator
-  #         file: ./abis/Aggregator.json
-  #     eventHandlers:
-  #       - event: 'AnswerUpdated(indexed int256,indexed uint256,uint256)'
-  #         handler: 'handleAggregatorAnswerUpdated'
-
   - name: PerpetualFuturesMarket
     kind: ethereum/contract
     source:
@@ -76,8 +36,7 @@ templates:
         - event: DelayedOrderSubmitted(indexed
             address,bool,int256,uint256,uint256,uint256,uint256,uint256,bytes32)
           handler: handleDelayedOrderSubmitted
-        - event: DelayedOrderRemoved(indexed
-            address,bool,uint256,int256,uint256,uint256,uint256,bytes32)
+        - event: DelayedOrderRemoved(indexed address,bool,uint256,int256,uint256,uint256,uint256,bytes32)
           handler: handleDelayedOrderRemoved
         - event: MarginTransferred(indexed address,int256)
           handler: handleMarginTransferred
