+++ v2/perps-v2/perps-subgraph/tests/perpsV2.test.ts
@@ -1,30 +1,19 @@
 import { Address, BigInt, Bytes, log, store } from '@graphprotocol/graph-ts';
-import {
-  assert,
-  describe,
-  test,
-  clearStore,
-  afterEach,
-  logStore,
-} from 'matchstick-as/assembly/index';
-import {
-  createDelayedOrderRemovedEvent,
-  createDelayedOrderSubmittedEvent,
-  createFunctionRecomputedEvent,
-  createMarginTransferredEvent,
-  createPositionLiquidatedEvent,
-  createPositionModifiedEvent,
-  toEth,
-  toGwei,
-} from './perpsV2-utils';
+import { afterEach, assert, clearStore, describe, test } from 'matchstick-as/assembly/index';
 import {
   handleDelayedOrderRemoved,
   handleDelayedOrderSubmitted,
   handleFundingRecomputed,
-  handleMarginTransferred,
-  handlePositionLiquidatedLegacy,
   handlePositionModified,
 } from '../src/futures';
+import {
+  createDelayedOrderRemovedEvent,
+  createDelayedOrderSubmittedEvent,
+  createFunctionRecomputedEvent,
+  createPositionModifiedEvent,
+  toEth,
+  toGwei,
+} from './perpsV2-utils';
 
 const trader = '0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045';
 
@@ -472,61 +461,6 @@ describe('Perps V2', () => {
     );
   });
 
-  test('position got liquidated', () => {
-    const positionOpenedEvent = createPositionModifiedEvent(
-      BigInt.fromI32(1),
-      Address.fromString(trader),
-      toEth(10),
-      toEth(1),
-      toEth(1),
-      toEth(1000),
-      BigInt.fromI32(1),
-      toGwei(2),
-      10,
-      BigInt.fromI32(12),
-      1
-    );
-    handlePositionModified(positionOpenedEvent);
-    const positionModifiedByLiquidationEvent = createPositionModifiedEvent(
-      BigInt.fromI32(1),
-      Address.fromString(trader),
-      toEth(0),
-      toEth(0),
-      toEth(0),
-      toEth(1000),
-      BigInt.fromI32(1),
-      toGwei(0),
-      10,
-      BigInt.fromI32(12),
-      2
-    );
-    handlePositionModified(positionModifiedByLiquidationEvent);
-    const positionLiquidatedEvent = createPositionLiquidatedEvent(
-      BigInt.fromI32(1),
-      Address.fromString(trader),
-      Address.fromString('0x98018d1eB5F2AE291D1537548ABBAA773382eEd4'),
-      toEth(10),
-      toEth(800),
-      toEth(1),
-      20,
-      3
-    );
-    handlePositionLiquidatedLegacy(positionLiquidatedEvent);
-    log.warning('STARTING ASSERTION', []);
-    // SYNTHETIX
-    log.info('Synthetix', []);
-    // 90% of the total fee
-    assert.fieldEquals('Synthetix', 'synthetix', 'feesByLiquidations', '849905364703000879');
-    // FUTURES TRADE
-    log.info('Futures Trade', []);
-    assert.fieldEquals(
-      'FuturesTrade',
-      `${positionLiquidatedEvent.transaction.hash.toHex()}-${2}`,
-      'type',
-      'Liquidated'
-    );
-  });
-
   test('open a short and close it', () => {
     const positionOpenedEvent = createPositionModifiedEvent(
       BigInt.fromI32(1),
@@ -614,110 +548,6 @@ describe('Perps V2', () => {
       'type',
       'PositionModified'
     );
-  });
-
-  test('Margin Transferred should update the position and produce an unknown tradeEntity', () => {
-    const initialFunding = 1000;
-    const initialFundingRecomputedEvent = createFunctionRecomputedEvent(
-      toEth(initialFunding), // funding
-      toEth(10), // fundingRate
-      BigInt.fromI32(1), // funding index
-      BigInt.fromI32(15), // block timestamp
-      10 // log index
-    );
-    handleFundingRecomputed(initialFundingRecomputedEvent);
-    const positionOpenedEvent = createPositionModifiedEvent(
-      BigInt.fromI32(1), //id
-      Address.fromString(trader), //account
-      toEth(500), // margin
-      toEth(-2), // size
-      toEth(-2), //tradeSize
-      toEth(1000), // lastPrice
-      BigInt.fromI32(1), // funding Index
-      toGwei(1000000000), // fee
-      10, // timestamp
-      BigInt.fromI32(12), // skew
-      1 // log index
-    );
-    handlePositionModified(positionOpenedEvent);
-    const positionId = `${positionOpenedEvent.address.toHex() + '-' + '0x1'}`;
-    assert.fieldEquals('FuturesPosition', positionId, 'realizedPnl', toEth(-1).toString());
-    assert.fieldEquals('FuturesPosition', positionId, 'margin', toEth(500).toString());
-    assert.fieldEquals('FuturesPosition', positionId, 'leverage', toEth(4).toString());
-    assert.fieldEquals('FuturesPosition', positionId, 'netFunding', toEth(0).toString());
-
-    const FundingRecomputedEvent1 = createFunctionRecomputedEvent(
-      toEth(initialFunding - 10), // funding going down
-      toEth(10), // fundingRate
-      BigInt.fromI32(2), //funding index
-      BigInt.fromI32(16), // block timestamp
-      10 // log index
-    );
-    handleFundingRecomputed(FundingRecomputedEvent1);
-    const marginTransferredEvent = createMarginTransferredEvent(
-      Address.fromString(trader), //sender
-      toEth(100), // marginDelta
-      10, // timestamp
-      1 // log index
-    );
-    handleMarginTransferred(marginTransferredEvent);
-    const positionUpdatedByTransferredMargin = createPositionModifiedEvent(
-      BigInt.fromI32(1), //id
-      Address.fromString(trader), //account
-      toEth(400), // margin
-      toEth(-2), // size
-      toEth(0), // tradeSize
-      toEth(1000), // lastPrice
-      BigInt.fromI32(2), // funding index
-      toGwei(0), // fee
-      10, // timestamp
-      BigInt.fromI32(12), // skew
-      2 // log index
-    );
-    handlePositionModified(positionUpdatedByTransferredMargin);
-    log.warning('STARTING ASSERTION', []);
-    log.info('Futures Trade', []);
-    assert.fieldEquals(
-      'FuturesTrade',
-      `${positionUpdatedByTransferredMargin.address.toHex()}-${BigInt.fromI32(1).toString()}`,
-      'type',
-      'PositionOpened'
-    );
-    assert.notInStore(
-      'FuturesTrade',
-      `${positionUpdatedByTransferredMargin.address.toHex()}-${BigInt.fromI32(2).toString()}`
-    );
-    assert.entityCount('FuturesTrade', 1);
-
-    assert.fieldEquals('FuturesPosition', positionId, 'netFunding', toEth(20).toString()); // (10 * 2)
-    assert.fieldEquals('FuturesPosition', positionId, 'realizedPnl', toEth(19).toString()); // -1 + (10 * 2)
-    assert.fieldEquals('FuturesPosition', positionId, 'margin', toEth(400).toString());
-    assert.fieldEquals('FuturesPosition', positionId, 'leverage', toEth(5).toString());
-
-    const FundingRecomputedEvent2 = createFunctionRecomputedEvent(
-      toEth(initialFunding + 10), // funding going down
-      toEth(10), // fundingRate
-      BigInt.fromI32(3), //funding index
-      BigInt.fromI32(16), // block timestamp
-      10 // log index
-    );
-    handleFundingRecomputed(FundingRecomputedEvent2);
-    const increasePos = createPositionModifiedEvent(
-      BigInt.fromI32(1), //id
-      Address.fromString(trader), //account
-      toEth(400), // margin
-      toEth(-3), // size
-      toEth(-1), // tradeSize
-      toEth(1000), // lastPrice
-      BigInt.fromI32(3), // funding index
-      toGwei(1000000000), // fee
-      10, // timestamp
-      BigInt.fromI32(12), // skew
-      2 // log index
-    );
-    handlePositionModified(increasePos);
-    assert.fieldEquals('FuturesPosition', positionId, 'netFunding', toEth(-20).toString()); // (10 * 2) + (-20* 2)
-    assert.fieldEquals('FuturesPosition', positionId, 'realizedPnl', toEth(-22).toString()); // -1 + (10 * 2) +  (-20 * 2) -1
   });
 
   test('funding recomputed', () => {
